Feature/async requests

+++ controllers/admin.js
@@ -83,6 +83,26 @@ exports.postAddProduct = (req, res, next) => {
     });
 };
 
+exports.addDummyProduct = (req, res, next) => {
+  const product = new Product({
+    title: 'Dummy Product',
+    price: 5,
+    description: 'This is my dummy product',
+    imageUrl: 'images\\dummy.jpg',
+    userId: req.user // Mongoose will pick the ID from the user object
+  });
+  product
+    .save()
+    .then(result => {
+      // console.log(result);
+      console.log('Created product!');
+      res.redirect('/admin/products');
+    })
+    .catch(err => {
+      console.log('Nope.');
+    });
+}
+
 exports.getEditProduct = (req, res, next) => { // ommit next if not using
   const editMode = req.query.edit;
   if(!editMode) {
@@ -166,24 +186,24 @@ exports.postEditProduct = (req, res, next) => {
     })
 }
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
   Product.findById(prodId)
   .then(product => {
     if(!product) {
       return next(new Error('Product not found.'));
     }
-    fileHelper.deleteFile(product.imageUrl);
+    if(product.imageUrl !== 'images\\dummy.jpg') {
+      fileHelper.deleteFile(product.imageUrl);
+    }
     return  Product.deleteOne({ _id: prodId, userId: req.user._id })
   })
   .then(() => {
     console.log('DESTROYED PRODUCT!');
-    res.redirect('/admin/products');
+    res.status(200).json({ message: 'Success!'});
   })
   .catch(err => {
-    const error = new Error(err);
-    error.httpStatusCode = 500;
-    return next(error);
+    res.status(500).json({ message: 'Delete failed!'});
   });
 }
